admin->delete nur aufrufen wenn auch admin gibt

+++ Vpc/Paragraphs/Row.php
@@ -4,6 +4,8 @@ class Vpc_Paragraphs_Row extends Vps_Db_Table_Row_Abstract
     protected function _delete()
     {
         $admin = Vpc_Admin::getInstance($this->component_class);
-        $admin->delete($this->page_id, $this->component_key . '-' . $this->id);
+        if ($admin) {
+            $admin->delete($this->page_id, $this->component_key . '-' . $this->id);
+        }
     }
 }
